Allow SelectRow to query a related table
